lab 2 finishied

+++ lab2/DebugPractice/DebugExercise1.java
@@ -4,8 +4,8 @@
  */
 public class DebugExercise1 {
     public static int divideThenRound(int top, int bottom) {
-        int quotient = top / bottom;
-        int result = Math.round(quotient);
+        double quotient = (double)top / (double)bottom;
+        int result = (int) Math.round(quotient);
         return result;
     }
 
@@ -24,5 +24,6 @@ public class DebugExercise1 {
         int b3 = 4;
         int result3 = divideThenRound(t3, b3);
         System.out.println("round(" + t3 + "/" + b3 + ")=" + result3);
+
     }
 }
